Add support for placements
