+++ api-report/purchases-js.api.json
@@ -2259,6 +2259,56 @@
                 "startIndex": 1,
                 "endIndex": 3
               }
+            },
+            {
+              "kind": "MethodSignature",
+              "canonicalReference": "@revenuecat/purchases-js!Offerings#getCurrentOfferingByPlacementId:member(1)",
+              "docComment": "/**\n * Retrieves a specific offering by a placement identifier. For more info see https://www.revenuecat.com/docs/tools/targeting\n *\n * @param placementId - The placement identifier to retrieve the offering for.\n */\n",
+              "excerptTokens": [
+                {
+                  "kind": "Content",
+                  "text": "getCurrentOfferingByPlacementId(placementId: "
+                },
+                {
+                  "kind": "Content",
+                  "text": "string"
+                },
+                {
+                  "kind": "Content",
+                  "text": "): "
+                },
+                {
+                  "kind": "Reference",
+                  "text": "Offering",
+                  "canonicalReference": "@revenuecat/purchases-js!Offering:interface"
+                },
+                {
+                  "kind": "Content",
+                  "text": " | null"
+                },
+                {
+                  "kind": "Content",
+                  "text": ";"
+                }
+              ],
+              "isOptional": false,
+              "returnTypeTokenRange": {
+                "startIndex": 3,
+                "endIndex": 5
+              },
+              "releaseTag": "Public",
+              "overloadIndex": 1,
+              "parameters": [
+                {
+                  "parameterName": "placementId",
+                  "parameterTypeTokenRange": {
+                    "startIndex": 1,
+                    "endIndex": 2
+                  },
+                  "isOptional": false
+                }
+              ],
+              "name": "getCurrentOfferingByPlacementId"
             }
           ],
           "extendsTokenRanges": []
@@ -2807,6 +2857,33 @@
               "isOptional": false,
               "releaseTag": "Public",
               "name": "offeringIdentifier",
+              "propertyTypeTokenRange": {
+                "startIndex": 1,
+                "endIndex": 2
+              }
+            },
+            {
+              "kind": "PropertySignature",
+              "canonicalReference": "@revenuecat/purchases-js!PresentedOfferingContext#placementIdentifier:member",
+              "docComment": "/**\n * If obtained this information from a placement, the identifier of the placement.\n */\n",
+              "excerptTokens": [
+                {
+                  "kind": "Content",
+                  "text": "readonly placementIdentifier: "
+                },
+                {
+                  "kind": "Content",
+                  "text": "string | null"
+                },
+                {
+                  "kind": "Content",
+                  "text": ";"
+                }
+              ],
+              "isReadonly": true,
+              "isOptional": false,
+              "releaseTag": "Public",
+              "name": "placementIdentifier",
               "propertyTypeTokenRange": {
                 "startIndex": 1,
                 "endIndex": 2
