+++ src/tests/main.test.ts
@@ -2,18 +2,10 @@ import { describe, expect, test } from "vitest";
 import {
   type CustomerInfo,
   type EntitlementInfo,
-  ErrorCode,
-  type Offering,
-  type Offerings,
-  type Package,
-  PackageType,
   Purchases,
   PurchasesError,
 } from "../main";
 import { UninitializedPurchasesError } from "../entities/errors";
-import { PeriodUnit } from "../helpers/duration-helper";
-import { createMonthlyPackageMock } from "./mocks/offering-mock-provider";
-import { failTest } from "./test-helpers";
 import {
   configurePurchases,
   testApiKey,
@@ -122,261 +114,6 @@ describe("Purchases.getAppUserId()", () => {
   test("returns app user id", () => {
     const purchases = configurePurchases();
     expect(purchases.getAppUserId()).toEqual(testUserId);
-  });
-});
-
-describe("getOfferings", () => {
-  const expectedMonthlyPackage = createMonthlyPackageMock();
-
-  test("can get offerings", async () => {
-    const purchases = configurePurchases();
-    const offerings = await purchases.getOfferings();
-
-    const currentOffering: Offering = {
-      serverDescription: "Offering 1",
-      identifier: "offering_1",
-      metadata: null,
-      packagesById: {
-        $rc_monthly: expectedMonthlyPackage,
-      },
-      availablePackages: [expectedMonthlyPackage],
-      lifetime: null,
-      annual: null,
-      sixMonth: null,
-      threeMonth: null,
-      twoMonth: null,
-      monthly: expectedMonthlyPackage,
-      weekly: null,
-    };
-
-    const subscriptionOption = {
-      id: "offer_12345",
-      priceId: "test_price_id",
-      base: {
-        cycleCount: 1,
-        periodDuration: "P1M",
-        period: {
-          number: 1,
-          unit: PeriodUnit.Month,
-        },
-        price: {
-          amount: 500,
-          amountMicros: 5000000,
-          currency: "USD",
-          formattedPrice: "$5.00",
-        },
-      },
-      trial: {
-        cycleCount: 1,
-        periodDuration: "P1W",
-        period: {
-          number: 1,
-          unit: PeriodUnit.Week,
-        },
-        price: null,
-      },
-    };
-
-    const package2: Package = {
-      identifier: "package_2",
-      packageType: PackageType.Custom,
-      rcBillingProduct: {
-        currentPrice: {
-          currency: "USD",
-          amount: 500,
-          amountMicros: 5000000,
-          formattedPrice: "$5.00",
-        },
-        displayName: "Monthly test 2",
-        title: "Monthly test 2",
-        description: "monthly description",
-        identifier: "monthly_2",
-        normalPeriodDuration: "P1M",
-        presentedOfferingIdentifier: "offering_2",
-        presentedOfferingContext: {
-          offeringIdentifier: "offering_2",
-          targetingContext: null,
-        },
-        defaultPurchaseOption: subscriptionOption,
-        defaultSubscriptionOption: subscriptionOption,
-        subscriptionOptions: {
-          offer_12345: subscriptionOption,
-        },
-      },
-    };
-
-    expect(offerings).toEqual({
-      all: {
-        offering_1: currentOffering,
-        offering_2: {
-          serverDescription: "Offering 2",
-          identifier: "offering_2",
-          metadata: null,
-          packagesById: {
-            package_2: package2,
-          },
-          availablePackages: [package2],
-          lifetime: null,
-          annual: null,
-          sixMonth: null,
-          threeMonth: null,
-          twoMonth: null,
-          monthly: null,
-          weekly: null,
-        },
-      },
-      current: currentOffering,
-    });
-  });
-
-  test("can get offerings without current offering id", async () => {
-    const purchases = configurePurchases("appUserIdWithoutCurrentOfferingId");
-    const offerings = await purchases.getOfferings();
-    const subscriptionOption = {
-      id: "offer_12345",
-      priceId: "test_price_id",
-      base: {
-        cycleCount: 1,
-        periodDuration: "P1M",
-        period: {
-          number: 1,
-          unit: PeriodUnit.Month,
-        },
-        price: {
-          amount: 500,
-          amountMicros: 5000000,
-          currency: "USD",
-          formattedPrice: "$5.00",
-        },
-      },
-      trial: {
-        cycleCount: 1,
-        periodDuration: "P1W",
-        period: {
-          number: 1,
-          unit: PeriodUnit.Week,
-        },
-        price: null,
-      },
-    };
-    const package2: Package = {
-      identifier: "package_2",
-      packageType: PackageType.Custom,
-      rcBillingProduct: {
-        currentPrice: {
-          currency: "USD",
-          amount: 500,
-          amountMicros: 5000000,
-          formattedPrice: "$5.00",
-        },
-        displayName: "Monthly test 2",
-        title: "Monthly test 2",
-        description: "monthly description",
-        identifier: "monthly_2",
-        normalPeriodDuration: "P1M",
-        presentedOfferingIdentifier: "offering_2",
-        presentedOfferingContext: {
-          offeringIdentifier: "offering_2",
-          targetingContext: null,
-        },
-        defaultPurchaseOption: subscriptionOption,
-        defaultSubscriptionOption: subscriptionOption,
-        subscriptionOptions: {
-          offer_12345: subscriptionOption,
-        },
-      },
-    };
-    const packageWithoutTargeting = createMonthlyPackageMock(null);
-    const expectedOfferings: Offerings = {
-      all: {
-        offering_1: {
-          serverDescription: "Offering 1",
-          identifier: "offering_1",
-          metadata: null,
-          packagesById: {
-            $rc_monthly: packageWithoutTargeting,
-          },
-          availablePackages: [packageWithoutTargeting],
-          lifetime: null,
-          annual: null,
-          sixMonth: null,
-          threeMonth: null,
-          twoMonth: null,
-          monthly: packageWithoutTargeting,
-          weekly: null,
-        },
-        offering_2: {
-          serverDescription: "Offering 2",
-          identifier: "offering_2",
-          metadata: null,
-          packagesById: {
-            package_2: package2,
-          },
-          availablePackages: [package2],
-          lifetime: null,
-          annual: null,
-          sixMonth: null,
-          threeMonth: null,
-          twoMonth: null,
-          monthly: null,
-          weekly: null,
-        },
-      },
-      current: null,
-    } as Offerings;
-
-    expect(offerings).toEqual(expectedOfferings);
-  });
-
-  test("can get offerings with missing products", async () => {
-    const purchases = configurePurchases("appUserIdWithMissingProducts");
-    const offerings = await purchases.getOfferings();
-
-    const packageWithoutTargeting = createMonthlyPackageMock(null);
-
-    const offering_1: Offering = {
-      serverDescription: "Offering 1",
-      identifier: "offering_1",
-      metadata: null,
-      packagesById: {
-        $rc_monthly: packageWithoutTargeting,
-      },
-      availablePackages: [packageWithoutTargeting],
-      lifetime: null,
-      annual: null,
-      sixMonth: null,
-      threeMonth: null,
-      twoMonth: null,
-      monthly: packageWithoutTargeting,
-      weekly: null,
-    };
-    expect(offerings).toEqual({
-      all: {
-        offering_1: offering_1,
-      },
-      current: null,
-    });
-  });
-
-  test("gets offerings with valid currency", async () => {
-    const purchases = configurePurchases();
-    await purchases.getOfferings({ currency: "EUR" }).then(
-      (offerings) => {
-        expect(offerings.current).not.toBeNull();
-      },
-      () => failTest(),
-    );
-  });
-
-  test("fails to get offerings with invalid currency", async () => {
-    const purchases = configurePurchases();
-    await purchases.getOfferings({ currency: "invalid" }).then(
-      () => failTest(),
-      (e) => {
-        expect(e).toBeInstanceOf(PurchasesError);
-        expect(e.errorCode).toEqual(ErrorCode.ConfigurationError);
-      },
-    );
   });
 });
