+++ src/tests/test-helpers.ts
@@ -1,8 +1,17 @@
 import { expect } from "vitest";
 import { PurchasesError } from "../entities/errors";
+import { type Offerings } from "../entities/offerings";
 
 export function failTest() {
   expect(true).toBeFalsy();
+}
+
+export function assertExpectedOfferings(
+  offerings: Offerings,
+  expectedOfferings: Offerings,
+) {
+  expect(offerings.all).toEqual(expectedOfferings.all);
+  expect(offerings.current).toEqual(expectedOfferings.current);
 }
 
 export function verifyExpectedError(e: unknown, expectedError: PurchasesError) {
